+++ src/redturtle/volto/testing.py
@@ -8,7 +8,7 @@ from plone.app.testing import PloneSandboxLayer
 from plone.restapi.testing import PloneRestApiDXLayer
 from plone.testing import z2
 
-import collective.volto.cookieconsent
+import collective.volto.gdprcookie
 import kitconcept.seo
 import plone.app.caching
 import plone.restapi
@@ -23,7 +23,7 @@ class RedturtleVoltoLayer(PloneSandboxLayer):
         # Load any other ZCML that is required for your tests.
         # The z3c.autoinclude feature is disabled in the Plone fixture base
         # layer.
-        self.loadZCML(package=collective.volto.cookieconsent)
+        self.loadZCML(package=collective.volto.gdprcookie)
         self.loadZCML(package=plone.restapi)
         self.loadZCML(package=redturtle.volto)
         self.loadZCML(package=plone.volto)
@@ -66,7 +66,7 @@ class RedturtleVoltoRestApiLayer(PloneRestApiDXLayer):
     def setUpZope(self, app, configurationContext):
         super(RedturtleVoltoRestApiLayer, self).setUpZope(app, configurationContext)
 
-        self.loadZCML(package=collective.volto.cookieconsent)
+        self.loadZCML(package=collective.volto.gdprcookie)
         self.loadZCML(package=plone.restapi)
         self.loadZCML(package=plone.volto)
         self.loadZCML(package=redturtle.volto)