+++ CHANGES.rst
@@ -8,7 +8,10 @@ Changelog
   [cekk]
 - Add flag in controlpanel to enable/disable INameChooser customization.
   [cekk]
-
+- Uninstall collective.volto.cookieconsent (deprecated). Will be removed from dependencies in next releases.
+  [cekk]
+- Add dependency to collective.volto.gdprcookie and install it by default.
+  [cekk]
 
 5.4.8 (2024-03-19)
 ------------------